refactor(test): switch to JS files instead of TS
